backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-generativeai==0.3.2
PyGithub==2.1.1
python-dotenv==1.0.0
pydantic==2.5.3
httpx==0.26.0
